Coerce to an EQ5D object

Generic for conversion to an EQ5D object. Methods are provided for data frame type objects.

The EQ5D class

We define an EQ5D object as a data frame that meets the following criteria:

It contains columns that represent dimensions from the EQ5D survey specification as well as a column representing the Visual Analogue Score.

It contains a column that acts as a unique respondent identifier and another that identifies different surveys over time. Together these should uniquely identify a response and no combination of these should be duplicated within the data frame.

Finally, it contains a column that provides the relative time of a response within the survey framework.

EQ5D3L, EQ5D5L and EQ5DY objects are defined as a subclass of EQ5D objects with the additional restriction that the corresponding dimension columns in x are either NA or whole numbers bounded below by 1 and above by 3 or 5 (depending on the survey type).

Usage

as_eq5d5l(x, ...)

# S3 method for default
as_eq5d5l(x, ...)

# S3 method for tbl_df
as_eq5d5l(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

# S3 method for data.table
as_eq5d5l(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

# S3 method for data.frame
as_eq5d5l(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

as_eq5d3l(x, ...)

# S3 method for default
as_eq5d3l(x, ...)

# S3 method for tbl_df
as_eq5d3l(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

# S3 method for data.table
as_eq5d3l(x, ...)

# S3 method for data.frame
as_eq5d3l(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

as_eq5dy(x, ...)

# S3 method for default
as_eq5dy(x, ...)

# S3 method for tbl_df
as_eq5dy(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

# S3 method for data.table
as_eq5dy(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

# S3 method for data.frame
as_eq5dy(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = vas,
  drop = FALSE,
  ...
)

Arguments

x

An R object.

...

Further arguments passed to or from other methods.

respondentID

[character]

The name of a variable in x that uniquely identifies respondents.

surveyID

[character]

Name of variable x that uniquely identifies surveys over time.

To avoid ambiguity the specified variable should be either numeric or a factor (in which case the order will be taken as that given by the factor levels).

A character variable in x will be accepted but converted, with warning, via as.factor().

time_index

Name of variable in x representing the relative time within the survey framework.

If the variable does not exist within x it will be created and set to NA_integer.

mobility

[character]

Name of the 'mobility' dimension in x.

self_care

[character]

Name of the 'self-care' dimension in x.

usual

[character]

Name of the 'usual activities' dimension in x.

pain

[character]

Name of the 'pain / discomfort' dimension in x.

anxiety

[character]

Name of the 'anxiety / depression' dimension in x.

vas

[character]

Name of the 'visual analogue score' variable in x.

drop

[logical]

Should additional columns, not specified by arguments, be dropped.

Value

An EQ5D5L, EQ5D3L or EQ5DY object.

See also

Examples

data("eq5d3l_example")
as_eq5d3l(eq5d3l_example)
#> # EQ-5D-3L: 200 x 10
#>    mobility self_care usual  pain anxiety Group  respondentID surveyID    vas
#>  *    <int>     <int> <int> <int>   <int> <chr>         <int>    <dbl>  <dbl>
#>  1        1         1     2     2       1 Group1            1        1 0.110 
#>  2        1         1     1     2       1 Group2            2        1 0.224 
#>  3        2         2     3     3       1 Group1            3        1 0.345 
#>  4        1         1     1     2       1 Group2            4        1 0.190 
#>  5        1         2     2     2       1 Group1            5        1 0.0892
#>  6        1         1     1     1       1 Group2            6        1 0.0592
#>  7        2         1     2     2       1 Group1            7        1 0.313 
#>  8        1         1     1     1       1 Group2            8        1 0.157 
#>  9        1         2     2     2       1 Group1            9        1 0.176 
#> 10        1         2     2     1       1 Group2           10        1 0.169 
#> # … with 190 more rows, and 1 more variable: time_index <int>